// src/app/api/models/stock-movement-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StockMovementResponseDto {
  createdAt: string;
  id: number;
  productId: number;
  productNameAtTime: string;
  quantity: number;
  reason?: string;
  type: 'in' | 'out' | 'adjustment' | 'usage';
}