// src/app/api/models/stock-summary-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StockSummaryResponseDto {
  availableStock: number;
  productCurrentPrice: number;
  productDescription?: string;
  productId: number;
  productName: string;
  productStatus: 'activo' | 'inactivo' | 'descatalogado' | 'agotado' | 'suspendido';
}